reject concurrent pre-receive hooks
Sandboxing the hook environment let multiple apps be deployed concurrently,
but git rejects commits if the target ref has been changed since the push began.

That can lead to a bad state where git-deploy hooks have modified the environment
but the commit is rejected.
This locking won't 100% prevent the bad state, but it reduces the window by the duration of
the pre-receive hook (i.e. <1 second instead of seconds->minutes)