PBL-21842: Provide encryption of environment variables
A common git-deploy pattern involves .env files that propogate through
etcd to environment variables for the target service.
If this pattern is used for passwords, it leaves a plaintext copy:
- in the git-deploy container
- in any clones of git-deploy repo/history
- in etcd (the scariest!)

This changeset introduces a new git-shell command: secret.
The new command generates a ciphertext that can be committed to git
in place of the original plaintext string.
This prevents plaintext storage in all of the above cases.

Presently all keys stay in the git-deploy container: export
and propagation are left as an exercise for downstream.
Eventually these could go be stored in etcd, but we need to
enable+trust etcd authentication before that becomes sane.